create data service interface
