+++ cdisc_rules_engine/interfaces/__init__.py
@@ -1,5 +1,6 @@
 from .cache_service_interface import CacheServiceInterface
 from .data_reader_interface import DataReaderInterface
+from .data_service_interface import DataServiceInterface
 from .factory_interface import FactoryInterface
 from .representation_interface import RepresentationInterface
 
@@ -7,6 +8,7 @@ from .representation_interface import RepresentationInterface
 __all__ = [
     "CacheServiceInterface",
     "DataReaderInterface",
+    "DataServiceInterface",
     "FactoryInterface",
     "RepresentationInterface",
 ]
