+++ cdisc_rules_engine/models/base_validation_entity.py
@@ -1,6 +1,8 @@
+from abc import ABC
+
 from cdisc_rules_engine.interfaces import RepresentationInterface
 
 
-class BaseValidationEntity(RepresentationInterface):
+class BaseValidationEntity(ABC, RepresentationInterface):
     def __init__(self):
         self.status = None
